try to resolve c# fail tests

+++ Assets/AltTester/Examples/Test/Editor/Driver/Testforscene1TestSample.cs
@@ -706,7 +706,7 @@ namespace Altom.AltDriver.Tests
         [Test]
         public void TestPressKeyWaitTheDuration()
         {
-            const int duration = 1;
+            const float duration = 1.0f;
             var button = altDriver.FindObject(By.NAME, "UIButton");
             altDriver.MoveMouse(button.GetScreenPosition());
             altDriver.PressKey(AltKeyCode.Mouse0, 1, duration);
